requirements.txt
instabot==0.117.0
pprint==0.1
python-dotenv==0.12.0